+++ src/main/java/kitchenpos/order/infra/DefaultKitchenridersClient.java
@@ -1,9 +1,8 @@
-package kitchenpos.infra;
-
-import org.springframework.stereotype.Component;
+package kitchenpos.order.infra;
 
 import java.math.BigDecimal;
 import java.util.UUID;
+import org.springframework.stereotype.Component;
 
 @Component
 public class DefaultKitchenridersClient implements KitchenridersClient {
