+++ src/test/java/kitchenpos/application/FakeKitchenridersClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.infra.KitchenridersClient;
+import kitchenpos.order.infra.KitchenridersClient;
 
 import java.math.BigDecimal;
 import java.util.UUID;
